fix storybook for whitelists
